Backend/requirements.txt
Flask==3.1.2
gunicorn==20.1.0
numpy==1.23.5
tensorflow-cpu==2.10.0
nltk==3.9.1
